Test case refactoring
