+++ conviva/src/main/java/com/bitmovin/analytics/conviva/ConvivaAnalytics.java
@@ -359,21 +359,22 @@ public class ConvivaAnalytics {
 
     private void createContentMetadata() {
         SourceItem sourceItem = bitmovinPlayer.getConfig().getSourceItem();
-
         if (sourceItem != null) {
             contentMetadataBuilder.setAssetName(sourceItem.getTitle());
         }
-
-        // Build custom tags
-        Map<String, String> customInternTags = new HashMap<>();
-        customInternTags.put("streamType", playerHelper.getStreamType());
-        customInternTags.put("integrationVersion", BuildConfig.VERSION_NAME);
-        contentMetadataBuilder.setCustom(customInternTags);
 
         this.buildDynamicContentMetadata();
     }
 
     private void buildDynamicContentMetadata() {
+        // Build custom tags here, though this is static metadata but
+        // streamType could be missing at time of session initialization
+        // as source information could be unavailable at that time
+        Map<String, String> customInternTags = new HashMap<>();
+        customInternTags.put("streamType", playerHelper.getStreamType());
+        customInternTags.put("integrationVersion", BuildConfig.VERSION_NAME);
+        contentMetadataBuilder.setCustom(customInternTags);
+
         if (bitmovinPlayer.isLive()) {
             contentMetadataBuilder.setStreamType(ContentMetadata.StreamType.LIVE);
         } else {
